# apps/server/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "user"
        ],
        "type": "object"
      },
...
      "CreateUserRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "JSON error response structure",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "EventDetailResponse": {
        "description": "Response for API (full detail)",
        "properties": {
          "data": {},
          "environment": {
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ingested_at": {
            "format": "date-time",
            "type": "string"
          },
          "issue_id": {
            "format": "uuid",
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "release": {
            "type": "string"
          },
          "sdk_name": {
            "type": "string"
          },
          "sdk_version": {
            "type": "string"
          },
          "server_name": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_id",
          "issue_id",
          "title",
          "timestamp",
          "ingested_at",
          "level",
          "platform",
          "release",
          "environment",
          "server_name",
          "sdk_name",
          "sdk_version",
          "data"
        ],
        "type": "object"
      },
      "EventResponse": {
        "description": "Response for API (list view)",
        "properties": {
          "environment": {
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "issue_id": {
            "format": "uuid",
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "release": {
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_id",
          "issue_id",
          "title",
          "timestamp",
          "level",
          "platform",
          "release",
          "environment"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "PaginatedEventResponse": {
        "description": "Paginated event response (cursor-based) - concrete type for OpenAPI",
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "type": "array"
          },
          "next_cursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "has_more"
        ],
        "type": "object"
      },
...
      "SortOrder": {
        "description": "Sort order direction",
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "is_admin": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_admin"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "bearerFormat": "hex",
        "description": "40-character hex API token",
        "scheme": "bearer",
        "type": "http"
      },
      "sentry_auth": {
        "description": "Sentry SDK authentication header or ?sentry_key= query parameter",
        "in": "header",
        "name": "X-Sentry-Auth",
        "type": "apiKey"
      },
      "session_auth": {
        "in": "cookie",
        "name": "rustrak_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Rustrak Team"
    },
    "description": "Ultra-lightweight error tracking server compatible with Sentry SDKs",
    "license": {
      "name": "GPL-3.0"
    },
    "title": "Rustrak API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/projects/{project_id}/issues/{issue_id}/events": {
      "get": {
        "operationId": "list_events",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Issue ID",
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Sort order direction (default: desc = newest first)",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          },
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEventResponse"
                }
              }
            },
            "description": "List of events"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Issue not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{project_id}/issues/{issue_id}/events\nLists events for an issue with cursor-based pagination",
        "tags": [
          "events"
        ]
      }
    },
    "/api/projects/{project_id}/issues/{issue_id}/events/{event_id}": {
      "get": {
        "operationId": "get_event",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Issue ID",
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Event ID",
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "Event details"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Event not found"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /api/projects/{project_id}/issues/{issue_id}/events/{event_id}\nGets a single event with full data",
        "tags": [
          "events"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Login successful"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid credentials"
          }
        },
        "summary": "POST /auth/login\nAuthenticate user and create session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Logged out"
          }
        },
        "security": [
          {
            "session_auth": []
          }
        ],
        "summary": "POST /auth/logout\nClear session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Current user"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "session_auth": []
          },
          {
            "bearer_auth": []
          }
        ],
        "summary": "GET /auth/me\nGet current authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "User created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "POST /auth/register\nCreate new user account",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer_auth": []
    },
    {
      "session_auth": []
    },
    {
      "sentry_auth": []
    }
  ],
...
}